+++ crates/cli/tests/run_test.rs
@@ -1034,28 +1034,6 @@ mod output_styles {
     }
 }
 
-mod multi_run {
-    use super::*;
-
-    #[test]
-    fn can_run_many_targets() {
-        let fixture = create_sandbox_with_git("cases");
-
-        let assert = create_moon_command(fixture.path())
-            .arg("run")
-            .arg("node:cjs")
-            .arg("node:mjs")
-            .assert();
-
-        let output = get_assert_output(&assert);
-
-        assert!(predicate::str::contains("node:cjs | stdout").eval(&output));
-        assert!(predicate::str::contains("node:mjs | stdout").eval(&output));
-        assert!(predicate::str::contains("node:cjs | stderr").eval(&output));
-        assert!(predicate::str::contains("node:mjs | stderr").eval(&output));
-    }
-}
-
 mod reports {
     use super::*;
